backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
openai==1.3.5
qdrant-client==1.8.0
python-multipart==0.0.6
better-exceptions==0.3.3
pytest==7.4.3
httpx==0.25.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
numpy==1.24.3
tiktoken==0.5.2